added firebase token support, added endpoint to send messages to users

+++ src/main/java/smartgate/capstonespringboot/security/JwtTokenProvider.java
@@ -25,7 +25,7 @@ public class JwtTokenProvider {
     private String jwtSecret;
 
     @Value("${app.jwtExpirationInMs}")
-    private int jwtExpirationInMs;
+    private Long jwtExpirationInMs;
 
     public String generateToken(Authentication authentication) {
